mcp_server/requirements.txt
model-context-protocol>=0.1.0
pymysql>=1.1.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0